Oppretter GrunnlagTjeneste og trekker ut metoder fra InntektsmeldingTjeneste som ikke handler om inntektsmelding
